test(e2e): apuntar el destino de la redirección en el rastro (#126)
El rastro que había apuntaba método, código y URL de la petición, y con eso
un fallo de #126 se leía «POST 303 → /panel/proveedores» — que se queda a
medias, porque esa URL es a dónde se ENVIÓ el POST, no a dónde decía la
respuesta que fuera.

El dato que falta para cerrar #126 es el `Location` de ese 303. Se venía
dando por hecho que sólo salía abriendo la traza del artefacto de CI, que
desde aquí no se puede descargar. No es cierto: la respuesta pasa por el
propio test y sus cabeceras están a mano.

Ahora se apunta, por cada POST:

  · si llevaba `next-action` — camino de cliente— o no —envío nativo;
  · el `location` / `x-action-redirect` de la respuesta.

Con eso las dos averías que hoy se ven idénticas se separan: si el destino
trae `?estado=` y el navegador no lo sigue, el problema está en el router de
cliente; si apunta a la misma URL sin query, está en el servidor.

De paso, el rastro sale de `panel-proveedores.spec.ts` a `utiles/rastro.ts`
—#126 asoma en cuatro specs y una copia por spec son cuatro sitios donde
arreglar lo mismo— y se cablea también en pagos y gastos.

Refs #126
