+++ tests/e2e/panel-gastos.spec.ts
@@ -3,6 +3,7 @@ import postgres from "postgres";
 
 import copy from "../../content/copy.es.json";
 import { RUTA_ACCESO, RUTA_GASTOS, RUTA_PANEL } from "../../src/config/constants";
+import { laPista, seguirLaPista } from "./utiles/rastro";
 
 /**
  * BODA-61 · Partidas de gasto
@@ -76,7 +77,8 @@ async function esperarEstado(pagina: Page, esperado: string) {
       .innerText()
       .catch(() => "(no se pudo leer la pantalla)");
     throw new Error(
-      `${(fallo as Error).message}\n\nLa pantalla decía:\n${enPantalla.slice(0, 600)}`,
+      `${(fallo as Error).message}\n\nLo que hizo la pestaña:\n${laPista(pagina)}` +
+        `\n\nLa pantalla decía:\n${enPantalla.slice(0, 600)}`,
     );
   }
 }
@@ -204,6 +206,7 @@ test.describe("Los gastos del presupuesto", () => {
     const categoria = await crearCategoria("Feliz");
     const concepto = `${MARCA} Ramo de novia`;
 
+    seguirLaPista(page);
     await entrar(page);
     await page.goto(RUTA_GASTOS);
